+++ lib/widgets/list/list_model.dart
@@ -6,6 +6,7 @@ import 'package:fml/log/manager.dart';
 import 'package:flutter/material.dart';
 import 'package:fml/widgets/form/form_model.dart';
 import 'package:fml/widgets/decorated/decorated_widget_model.dart';
+import 'package:fml/widgets/selectable/selectable_model.dart';
 import 'package:xml/xml.dart';
 import 'package:fml/event/handler.dart'            ;
 import 'package:fml/widgets/list/list_view.dart';
@@ -14,7 +15,7 @@ import 'package:fml/widgets/widget/widget_model.dart'     ;
 import 'package:fml/observable/observable_barrel.dart';
 import 'package:fml/helper/common_helpers.dart';
 
-class ListModel extends DecoratedWidgetModel implements IForm, IScrolling
+class ListModel extends SelectableModel implements IForm, IScrolling
 {
   final HashMap<int,ListItemModel> items = HashMap<int,ListItemModel>();
 
@@ -34,36 +35,6 @@ class ListModel extends DecoratedWidgetModel implements IForm, IScrolling
     }
   }
   bool get scrollShadows => _scrollShadows?.get() ?? false;
-
-  IntegerObservable? _selectedIndex;
-  set selectedIndex (dynamic v)
-  {
-    if (_selectedIndex != null)
-    {
-      _selectedIndex!.set(v);
-    }
-    else if (v != null)
-    {
-      _selectedIndex = IntegerObservable(Binding.toKey(id, 'selectedindex'), v, scope: scope);
-    }
-  }
-  int? get selectedIndex => _selectedIndex?.get();
-
-  // selected selected element
-  ListObservable? _selected;
-  set selected(dynamic v)
-  {
-    if (_selected != null)
-    {
-      _selected!.set(v);
-    }
-    else if (v != null)
-    {
-      _selected = ListObservable(Binding.toKey(id, 'selected'), null, scope: scope);
-      _selected!.set(v);
-    }
-  }
-  get selected => _selected?.get();
   
   BooleanObservable? _scrollButtons;
   set scrollButtons (dynamic v)
